refactored code by moving grammar into a own package
