+++ src/main/kotlin/isml/aidev/SymbolsNode.kt
@@ -1,5 +1,6 @@
 package isml.aidev
 
+import isml.aidev.grammar.Symbol
 import isml.aidev.util.Evaluator
 
 /**
