+++ src/main/kotlin/isml/aidev/starlib/PCFGSearchInput.kt
@@ -1,8 +1,8 @@
 package isml.aidev.starlib
 
-import isml.aidev.Grammar
 import isml.aidev.RuleEdge
 import isml.aidev.SymbolsNode
+import isml.aidev.grammar.Grammar
 import org.api4.java.ai.graphsearch.problem.IPathSearchInput
 import org.api4.java.ai.graphsearch.problem.implicit.graphgenerator.IPathGoalTester
